+++ extensions/rnuilib-snippets/snippets/snippets.code-snippets
@@ -253,9 +253,9 @@
   },
   "Marquee": {
     "prefix": "marquee",
-    "description": "Marquee component for sliding text",
+    "description": "Marquee component for sliding text (Incubator)",
     "body": [
-      "<Marquee",
+      "<Incubator.Marquee",
       "  key={`${index}`$1}",
       "  label={'Hey there'$2}",
       "  direction={directionHorizontal ? MarqueeDirections.LEFT : MarqueeDirections.RIGHT$3}",
