+++ packages/react-native-ui-lib/src/incubator/marquee/marquee.api.json
@@ -2,7 +2,7 @@
   "name": "Marquee",
   "category": "media",
   "description": "Marquee component for sliding text",
-  "example": "https://github.com/wix/react-native-ui-lib/blob/master/demo/src/screens/componentScreens/MarqueeScreen.tsx",
+  "example": "https://github.com/wix/react-native-ui-lib/blob/master/demo/src/screens/incubatorScreens/IncubatorMarqueeScreen.tsx",
   "props": [
     {"name": "label", "type": "string", "description": "Marquee label"},
     {"name": "labelStyle", "type": "TextProps['style']", "description": "Marquee label style"},
